+++ packages/liveblocks-python/liveblocks/api/comments/subscribe_to_thread.py
@@ -1,0 +1,80 @@
+from typing import Any
+from urllib.parse import quote
+
+import httpx
+
+from ... import errors
+from ...models.subscribe_to_thread_request_body import SubscribeToThreadRequestBody
+from ...models.subscription import Subscription
+
+
+def _get_kwargs(
+    room_id: str,
+    thread_id: str,
+    *,
+    body: SubscribeToThreadRequestBody,
+) -> dict[str, Any]:
+    headers: dict[str, Any] = {}
+
+    _kwargs: dict[str, Any] = {
+        "method": "post",
+        "url": "/v2/rooms/{room_id}/threads/{thread_id}/subscribe".format(
+            room_id=quote(str(room_id), safe=""),
+            thread_id=quote(str(thread_id), safe=""),
+        ),
+    }
+
+    _kwargs["json"] = body.to_dict()
+
+    headers["Content-Type"] = "application/json"
+
+    _kwargs["headers"] = headers
+    return _kwargs
+
+
+def _parse_response(*, response: httpx.Response) -> Subscription:
+    if response.status_code == 200:
+        response_200 = Subscription.from_dict(response.json())
+
+        return response_200
+
+    raise errors.LiveblocksError.from_response(response)
+
+
+def _sync(
+    room_id: str,
+    thread_id: str,
+    *,
+    client: httpx.Client,
+    body: SubscribeToThreadRequestBody,
+) -> Subscription:
+    kwargs = _get_kwargs(
+        room_id=room_id,
+        thread_id=thread_id,
+        body=body,
+    )
+
+    response = client.request(
+        **kwargs,
+    )
+    return _parse_response(response=response)
+
+
+async def _asyncio(
+    room_id: str,
+    thread_id: str,
+    *,
+    client: httpx.AsyncClient,
+    body: SubscribeToThreadRequestBody,
+) -> Subscription:
+    kwargs = _get_kwargs(
+        room_id=room_id,
+        thread_id=thread_id,
+        body=body,
+    )
+
+    response = await client.request(
+        **kwargs,
+    )
+
+    return _parse_response(response=response)