+++ packages/liveblocks-python/liveblocks/models/room_permission_item.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+
+class RoomPermissionItem(StrEnum):
+    COMMENTSWRITE = "comments:write"
+    ROOMPRESENCEWRITE = "room:presence:write"
+    ROOMREAD = "room:read"
+    ROOMWRITE = "room:write"